refactor (transcription): move time normalizer into normalization and split laughter pass
- move `TimeExpressionNormalizer.swift` under `Core/Transcription/Normalization` and update Xcode project wiring
- add `TranscriptionLaughterNormalizer.swift` to isolate laughter cleanup from time normalization
- update `TranscriptionPostProcessor` to use the extracted laughter normalizer in the pipeline
- keep post-processor focused on orchestration while preserving existing normalization behavior
- update `Docs/CODEMAP.md` and `Docs/ENGINEERING.md` for the new normalization structure
- verify with clean focused tests (`DictationPipelineSmokeTests`, `TranscriptionPostProcessorTests`) passing
