+++ Core/Transcription/TranscriptionPostProcessor.swift
@@ -2,20 +2,15 @@ import Foundation
 
 @MainActor
 final class TranscriptionPostProcessor {
-    private static let domainLikeTokenRegex: NSRegularExpression? = try? NSRegularExpression(
-        pattern: #"(?i)^(?:https?://)?(?:www\.)?[a-z0-9\-]+(?:\.[a-z0-9\-]+)+/?$"#,
-        options: []
-    )
-    private static let commonTopLevelDomains: Set<String> = [
-        "com", "net", "org", "io", "app", "dev", "ai", "co", "me", "edu", "gov",
-        "us", "uk", "ca", "au", "de", "fr", "jp"
-    ]
-
     private let enablePhoneticMatcher = true
     private let vocabularyNormalizer = CustomVocabularyNormalizer()
     private let dictionaryMatcher = DictionaryMatcher()
     private let listFormattingEngine = ListFormattingEngine()
+    private let laughterNormalizer = TranscriptionLaughterNormalizer()
     private let timeExpressionNormalizer = TimeExpressionNormalizer()
+    private let whitespaceNormalizer = TranscriptionWhitespaceNormalizer()
+    private let capitalizationNormalizer = TranscriptionCapitalizationNormalizer()
+    private let terminalPunctuationNormalizer = TranscriptionTerminalPunctuationNormalizer()
     private var dictionaryFingerprint = ""
 
     // Keep teardown executor-agnostic to avoid runtime deinit crashes in test host.
@@ -82,7 +77,7 @@ final class TranscriptionPostProcessor {
         #if DEBUG
         logPipelineStage("listFormatted", listFormatted)
         #endif
-        let laughterNormalized = normalizeLaughterExpressions(in: listFormatted)
+        let laughterNormalized = laughterNormalizer.normalize(in: listFormatted)
         #if DEBUG
         logPipelineStage("laughterNormalized", laughterNormalized)
         #endif
@@ -94,285 +89,13 @@ final class TranscriptionPostProcessor {
         #if DEBUG
         logPipelineStage("emailNormalizedOutput", emailNormalizedOutput)
         #endif
-        let whitespaceNormalized = normalizeOutputWhitespace(emailNormalizedOutput, renderMode: renderMode)
-        let textStartNormalized = capitalizeAtTextStart(whitespaceNormalized)
-        let sentenceStartNormalized = capitalizeAfterSentenceBoundary(textStartNormalized)
-        let lineStartNormalized = capitalizeAfterLineBreak(sentenceStartNormalized)
-        let output = appendTerminalPeriodIfEndingInFormattedTime(lineStartNormalized)
+        let whitespaceNormalized = whitespaceNormalizer.normalize(emailNormalizedOutput, renderMode: renderMode)
+        let sentenceNormalized = capitalizationNormalizer.normalizeSentenceStarts(in: whitespaceNormalized)
+        let output = terminalPunctuationNormalizer.appendTerminalPeriodIfEndingInFormattedTime(sentenceNormalized)
         #if DEBUG
         logPipelineStage("output", output)
         #endif
         return output
-    }
-
-    private func normalizeOutputWhitespace(_ text: String, renderMode: ListRenderMode) -> String {
-        switch renderMode {
-        case .singleLineInline:
-            return text
-                .replacingOccurrences(of: "\\s+", with: " ", options: .regularExpression)
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-        case .multiline:
-            let normalizedLines = text
-                .split(separator: "\n", omittingEmptySubsequences: false)
-                .map { line in
-                    String(line)
-                        .replacingOccurrences(of: "[\\t ]+", with: " ", options: .regularExpression)
-                        .trimmingCharacters(in: .whitespacesAndNewlines)
-                }
-
-            var collapsedLines: [String] = []
-            collapsedLines.reserveCapacity(normalizedLines.count)
-
-            var previousWasBlank = false
-            for line in normalizedLines {
-                let isBlank = line.isEmpty
-                if isBlank {
-                    if collapsedLines.isEmpty || previousWasBlank {
-                        continue
-                    }
-                    collapsedLines.append("")
-                    previousWasBlank = true
-                } else {
-                    collapsedLines.append(line)
-                    previousWasBlank = false
-                }
-            }
-
-            while collapsedLines.first?.isEmpty == true {
-                collapsedLines.removeFirst()
-            }
-            while collapsedLines.last?.isEmpty == true {
-                collapsedLines.removeLast()
-            }
-
-            return collapsedLines.joined(separator: "\n")
-        }
-    }
-
-    private func appendTerminalPeriodIfEndingInFormattedTime(_ text: String) -> String {
-        guard !text.isEmpty else { return text }
-
-        // Respect existing terminal punctuation, including punctuation before closing quotes/brackets.
-        if text.range(of: #"[.!?…][\"'”’\)\]\}]*\s*$"#, options: .regularExpression) != nil {
-            return text
-        }
-
-        let terminalTimePattern = #"(?i)\b(?:[1-9]|1[0-2]):[0-5][0-9]\s(?:AM|PM)\s*$"#
-        guard let regex = try? NSRegularExpression(pattern: terminalTimePattern) else { return text }
-
-        let nsText = text as NSString
-        let range = NSRange(location: 0, length: nsText.length)
-        guard let match = regex.firstMatch(in: text, options: [], range: range) else {
-            return text
-        }
-
-        // Only treat this as sentence-like if there is prose before the terminal time.
-        let prefix = nsText.substring(to: match.range.location)
-        guard prefix.range(of: #"\b[A-Za-z]{3,}\b"#, options: .regularExpression) != nil else {
-            return text
-        }
-
-        return text + "."
-    }
-
-    private func capitalizeAtTextStart(_ text: String) -> String {
-        guard !text.isEmpty else { return text }
-        guard let regex = try? NSRegularExpression(
-            pattern: #"^(\s*["'“”‘’\(\[\{]*)([a-z])"#,
-            options: []
-        ) else {
-            return text
-        }
-
-        let nsText = text as NSString
-        let range = NSRange(location: 0, length: nsText.length)
-        guard let match = regex.firstMatch(in: text, options: [], range: range) else { return text }
-        if isAddressOrURLToken(firstToken(in: text)) { return text }
-
-        let prefix = nsText.substring(with: match.range(at: 1))
-        let nextLetter = nsText.substring(with: match.range(at: 2)).uppercased()
-        let mutable = NSMutableString(string: text)
-        mutable.replaceCharacters(in: match.range, with: "\(prefix)\(nextLetter)")
-        return mutable as String
-    }
-
-    private func capitalizeAfterSentenceBoundary(_ text: String) -> String {
-        guard !text.isEmpty else { return text }
-        guard let boundaryRegex = try? NSRegularExpression(
-            pattern: #"(?<!\d)([.!?;:…]["'”’\)\]\}]*)(\s*)([a-z])"#,
-            options: []
-        ) else {
-            return text
-        }
-
-        let nsText = text as NSString
-        let fullRange = NSRange(location: 0, length: nsText.length)
-        let matches = boundaryRegex.matches(in: text, options: [], range: fullRange)
-        guard !matches.isEmpty else { return text }
-
-        let mutable = NSMutableString(string: text)
-        for match in matches.reversed() {
-            let tokenStart = match.range(at: 3).location
-            let tokenTail = nsText.substring(with: NSRange(location: tokenStart, length: nsText.length - tokenStart))
-            let firstToken = tokenTail.split(whereSeparator: \.isWhitespace).first.map(String.init) ?? ""
-            if isAddressOrURLToken(firstToken) {
-                continue
-            }
-
-            let boundaryText = nsText.substring(with: match.range(at: 1))
-            if boundaryText.first == "." {
-                if isLikelyDomainBoundary(text, dotLocation: match.range(at: 1).location) {
-                    continue
-                }
-
-                let prefixText = nsText.substring(to: match.range(at: 1).location)
-                let previousToken = prefixText.split(whereSeparator: \.isWhitespace).last.map(String.init) ?? ""
-                if previousToken.contains("@") {
-                    continue
-                }
-            }
-
-            let prefix = nsText.substring(with: match.range(at: 1))
-            let spacing = nsText.substring(with: match.range(at: 2))
-            let separator = spacing.isEmpty ? " " : spacing
-            let nextLetter = nsText.substring(with: match.range(at: 3)).uppercased()
-            mutable.replaceCharacters(in: match.range, with: "\(prefix)\(separator)\(nextLetter)")
-        }
-
-        return mutable as String
-    }
-
-    private func capitalizeAfterLineBreak(_ text: String) -> String {
-        guard !text.isEmpty else { return text }
-        guard let lineBreakRegex = try? NSRegularExpression(
-            pattern: #"(\n)([ \t]*["'“”‘’\(\[\{]*)([a-z])"#,
-            options: []
-        ) else {
-            return text
-        }
-
-        let nsText = text as NSString
-        let fullRange = NSRange(location: 0, length: nsText.length)
-        let matches = lineBreakRegex.matches(in: text, options: [], range: fullRange)
-        guard !matches.isEmpty else { return text }
-
-        let mutable = NSMutableString(string: text)
-        for match in matches.reversed() {
-            let tokenStart = match.range(at: 3).location
-            guard !lineStartsWithAddressOrURL(in: nsText, from: tokenStart) else { continue }
-            let newline = nsText.substring(with: match.range(at: 1))
-            let prefix = nsText.substring(with: match.range(at: 2))
-            let nextLetter = nsText.substring(with: match.range(at: 3)).uppercased()
-            mutable.replaceCharacters(in: match.range, with: "\(newline)\(prefix)\(nextLetter)")
-        }
-
-        return mutable as String
-    }
-
-    private func lineStartsWithAddressOrURL(in text: NSString, from location: Int) -> Bool {
-        guard location >= 0, location < text.length else { return false }
-        let suffix = text.substring(with: NSRange(location: location, length: text.length - location))
-        let line = suffix.split(separator: "\n", maxSplits: 1, omittingEmptySubsequences: false).first.map(String.init) ?? ""
-        let firstToken = firstToken(in: line)
-        guard !firstToken.isEmpty else { return false }
-        return isAddressOrURLToken(firstToken)
-    }
-
-    private func isAddressOrURLToken(_ token: String) -> Bool {
-        let stripped = token.trimmingCharacters(in: CharacterSet(charactersIn: "\"'“”‘’()[]{}.,;:!?"))
-        guard !stripped.isEmpty else { return false }
-
-        let lowered = stripped.lowercased()
-        if lowered.contains("@") {
-            let parts = lowered.split(separator: "@", maxSplits: 1, omittingEmptySubsequences: false)
-            if parts.count == 2, !parts[0].isEmpty, parts[1].contains(".") {
-                return true
-            }
-        }
-
-        var candidate = lowered
-        if candidate.hasPrefix("http://") {
-            candidate.removeFirst("http://".count)
-        } else if candidate.hasPrefix("https://") {
-            candidate.removeFirst("https://".count)
-        }
-        if let hostOnly = candidate.split(separator: "/", maxSplits: 1, omittingEmptySubsequences: false).first {
-            candidate = String(hostOnly)
-        }
-
-        let range = NSRange(location: 0, length: (candidate as NSString).length)
-        if let domainRegex = Self.domainLikeTokenRegex,
-           domainRegex.firstMatch(in: candidate, options: [], range: range) != nil {
-            return true
-        }
-        return false
-    }
-
-    private func firstToken(in text: String) -> String {
-        text
-            .split(whereSeparator: \.isWhitespace)
-            .first
-            .map(String.init) ?? ""
-    }
-
-    private func isLikelyDomainBoundary(_ text: String, dotLocation: Int) -> Bool {
-        guard dotLocation >= 0 else { return false }
-        let nsText = text as NSString
-        guard dotLocation < nsText.length else { return false }
-        guard dotLocation + 1 < nsText.length else { return false }
-
-        let domainLabelCharacterSet = CharacterSet(charactersIn: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-")
-        guard let nextScalar = UnicodeScalar(nsText.character(at: dotLocation + 1)),
-              domainLabelCharacterSet.contains(nextScalar) else {
-            return false
-        }
-
-        var start = dotLocation
-        while start > 0 {
-            guard let scalar = UnicodeScalar(nsText.character(at: start - 1)) else { break }
-            if CharacterSet.whitespacesAndNewlines.contains(scalar) {
-                break
-            }
-            start -= 1
-        }
-
-        var end = dotLocation + 1
-        while end < nsText.length {
-            guard let scalar = UnicodeScalar(nsText.character(at: end)) else { break }
-            if CharacterSet.whitespacesAndNewlines.contains(scalar) {
-                break
-            }
-            end += 1
-        }
-
-        let tokenRange = NSRange(location: start, length: end - start)
-        guard tokenRange.length > 0 else { return false }
-
-        let token = nsText.substring(with: tokenRange)
-            .trimmingCharacters(in: CharacterSet(charactersIn: "\"'“”‘’()[]{}"))
-            .replacingOccurrences(
-                of: #"[.,;:!?]+$"#,
-                with: "",
-                options: .regularExpression
-            )
-        guard !token.isEmpty, !token.contains("@"), token.contains(".") else { return false }
-
-        guard let regex = Self.domainLikeTokenRegex else { return false }
-        let fullRange = NSRange(location: 0, length: (token as NSString).length)
-        guard regex.firstMatch(in: token, options: [], range: fullRange) != nil else { return false }
-
-        let normalized = token.lowercased()
-        if normalized.hasPrefix("http://") || normalized.hasPrefix("https://") || normalized.hasPrefix("www.") {
-            return true
-        }
-
-        let dotCount = normalized.filter { $0 == "." }.count
-        if dotCount >= 2 {
-            return true
-        }
-
-        guard let tld = normalized.split(separator: ".").last.map(String.init) else { return false }
-        return Self.commonTopLevelDomains.contains(tld)
     }
 
     private func fingerprint(for entries: [DictionaryEntry]) -> String {
@@ -380,12 +103,6 @@ final class TranscriptionPostProcessor {
             .map { "\($0.id.uuidString):\($0.phrase)" }
             .sorted()
             .joined(separator: "|")
-    }
-
-    private func normalizeLaughterExpressions(in text: String) -> String {
-        replacingMatches(pattern: #"\bha\s+ha\b"#, in: text) { _, _ in
-            "haha"
-        }
     }
 
     private func normalizeIdioms(in text: String) -> String {
